Implemented daily_email_functionality as body text only issue #65
