+++ app/wsmodules/df_cleaner.py
@@ -63,6 +63,52 @@ def clean_sqm_eur_col(df_name):
     return final_df
 
 
+def save_text_report_to_file(text: list, file_name: str) -> None:
+    """Writes oneline data text to mailer report file"""
+    with open(file_name, 'a') as the_file:
+        for line in text:
+            the_file.write(f"{line}\n")
+
+
+def create_email_body(clean_data_frame, file_name: str) -> None:
+    """Creates categorized by room count ad hash : data for email body.
+
+    Requires:
+        clean_data_frame: pandas data frame
+
+    Creates:
+        email_body_txt_m4.txt : text file"""
+
+    email_body_txt = ['Sendgrid mailer Milestone 4 report:']
+    for room_count in range(4):
+        room_count_str = str(room_count + 1)
+        section_line = str(room_count_str + " room apartment section: ")
+        email_body_txt.append(section_line)
+        filtered_by_room_count = clean_data_frame.loc[clean_data_frame['Room_count'] == int(room_count_str)]
+        colum_line = "[Rooms, Floor, Size , Price, SQM Price, Apartment Street, Pub_date,  URL]"
+        email_body_txt.append(colum_line)
+        for index, row in filtered_by_room_count.iterrows():
+            url_str = row["URL"]
+            sqm_str = row["Size_sqm"]
+            floor_str = row["Floor"]
+            total_price = row["Price_in_eur"]
+            sqm_price = row['SQ_meter_price']
+            rooms_str = row['Room_count']
+            street_str = row['Street']
+            pub_date_str = row['Pub_date']
+            report_line = "  " + str(rooms_str) + "     " + \
+                          str(floor_str) + "    " + \
+                          str(sqm_str) + "   " + \
+                          str(total_price) + "    " + \
+                          str(sqm_price) + "   " + \
+                          str(street_str) + "   " + \
+                          str(pub_date_str) + " " + \
+                          str(url_str)
+            email_body_txt.append(report_line)
+        save_text_report_to_file(email_body_txt, file_name)
+
+
+
 def df_cleaner_main():
     """ Cleans df, sorts df by price in EUR, save to csv file """
     clean_df = clean_data_frame(df_to_clean)
@@ -71,7 +117,10 @@ def df_cleaner_main():
     clean_df = clean_sqm_eur_col(clean_price_col)
     sorted_df = clean_df.sort_values(by='Price_in_eur', ascending=True)
     sorted_df.to_csv("cleaned-sorted-df.csv")
+    all_ads_df = pd.read_csv("cleaned-sorted-df.csv", index_col=False)
+    create_email_body(all_ads_df, 'email_body_txt_m4.txt')
     print("Debug info: Completed dat_formater module ... ")
+
 
 
 # Main module code driver
